Moved to one main job class.
